Reorganized files, starting to work on a script to carry out the entire workflow
