--- backend/scripts/workflow.js
@@ -1,1 +1,0 @@
-export const votingWorkflow = () => {};
